refactor : InputView에서 반환하는 자료형 변경
