+++ src/test/java/domain/TicketTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
@@ -55,25 +56,22 @@ public class TicketTest {
 
     @Test
     void 입력받은_문자열로_로또번호_정상생성() {
-        assertThatCode(() -> Ticket.from("1, 2, 3, 4, 5, 6"))
+        assertThatCode(() -> Ticket.from(List.of(1, 2, 3, 4, 5, 6)))
                 .doesNotThrowAnyException();
     }
 
     @ParameterizedTest(name = "{index}: {1}")
     @MethodSource("parameters")
     @DisplayName("로또 번호 입력 테스트")
-    void 로또번호입력(String input, String testName) {
+    void 로또번호입력(List<Integer> input, String testName) {
         assertThatThrownBy(() -> Ticket.from(input))
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
     static Stream<Arguments> parameters() {
         return Stream.of(
-                Arguments.of("1, 2, 3, 4, 5", "5개입력"),
-                Arguments.of("1, 2, 3, 4, 5, 6, 7", "7개입력"),
-                Arguments.of("1, 2, 3, 4, 5, a", "문자 개입력"),
-                Arguments.of("1, 2, 3, 4, 5, 5", "중복 번호 입력"),
-                Arguments.of("1, 2, 3, 4, 5, 6, ", "마지막에 구분자 추가"),
-                Arguments.of(null, "널 입력"));
+                Arguments.of(List.of(1, 2, 3, 4, 5), "5개입력"),
+                Arguments.of(List.of(1, 2, 3, 4, 5, 6, 7), "7개입력"),
+                Arguments.of(List.of(1, 2, 3, 4, 5, 5), "중복 번호 입력"));
     }
 }
